rename test utils file
