+++ frontend/src/components/accounts/__tests__/AccountDetails.test.tsx
@@ -1,4 +1,4 @@
-import { renderI18nElement } from "../../../libraries/test";
+import { renderI18nElement } from "../../../libraries/tester";
 
 import AccountDetails from "../AccountDetails";
