+++ frontend/src/components/blocks/__tests__/BlockDetails.test.tsx
@@ -1,6 +1,6 @@
 import BN from "bn.js";
 
-import { renderI18nElement } from "../../../libraries/test";
+import { renderI18nElement } from "../../../libraries/tester";
 
 import BlockDetails from "../BlockDetails";
