+++ frontend/src/components/blocks/__tests__/BlocksRow.test.tsx
@@ -1,4 +1,4 @@
-import { renderI18nElement } from "../../../libraries/test";
+import { renderI18nElement } from "../../../libraries/tester";
 
 import BlocksRow from "../BlocksRow";
