+++ frontend/src/components/nodes/__test__/NodeRow.test.tsx
@@ -1,4 +1,4 @@
-import { renderI18nElement } from "../../../libraries/test";
+import { renderI18nElement } from "../../../libraries/tester";
 
 import NodeRow from "../NodeRow";
