+++ frontend/src/components/transactions/__tests__/ReceiptRow.test.tsx
@@ -1,4 +1,4 @@
-import { renderI18nElement } from "../../../libraries/test";
+import { renderI18nElement } from "../../../libraries/tester";
 
 import ReceiptRow from "../ReceiptRow";
