+++ frontend/src/components/utils/__tests__/AccountLink.test.tsx
@@ -1,4 +1,4 @@
-import { renderI18nElement } from "../../../libraries/test";
+import { renderI18nElement } from "../../../libraries/tester";
 
 import AccountLink from "../AccountLink";
