+++ frontend/src/components/utils/__tests__/ReceiptLink.test.tsx
@@ -1,4 +1,4 @@
-import { renderI18nElement } from "../../../libraries/test";
+import { renderI18nElement } from "../../../libraries/tester";
 
 import ReceiptLink from "../ReceiptLink";
