+++ frontend/src/components/utils/__tests__/TransactionLink.test.tsx
@@ -1,4 +1,4 @@
-import { renderI18nElement } from "../../../libraries/test";
+import { renderI18nElement } from "../../../libraries/tester";
 
 import TransactionLink from "../TransactionLink";
